edited bash file

+++ OpenAI-STT-TTS/deploy.sh
@@ -1,41 +1,31 @@
-#!/bin/bash
-
 # Define symlinks
-# Check if version.tf does not exist
-if [ ! -e "version.tf" ]; then
-    ln -s ../version.tf version.tf
-fi
+[[ ! -e version.tf ]] && ln -s ../version.tf version.tf
+[[ ! -e providers.tf ]] && ln -s ../providers.tf providers.tf
+[[ ! -e resource-group.tf ]] && ln -s ../resource-group.tf resource-group.tf
 
-# Check if providers.tf does not exist
-if [ ! -e "providers.tf" ]; then
-    ln -s ../providers.tf providers.tf
-fi
-
-# Check if resource-group.tf does not exist
-if [ ! -e "resource-group.tf" ]; then
-    ln -s ../resource-group.tf resource-group.tf
-fi
-
-# Run terraform init
+# Run terraform commands
+terraform fmt
 terraform init
-
-# Run terraform plan
 terraform plan
-
-# Run terraform apply
 terraform apply -auto-approve
 
+# Path to the .env file in the child directory
+envFilePath="./src/.env"
+
 # Clear the .env file
-if [ -e ".env" ]; then
-    rm .env
-fi
+[[ -e $envFilePath ]] && rm $envFilePath
 
 # Get the output and save to .env file
-# Define an array of output keys
-outputKeys=("OPEN_AI_ENDPOINT" "OPEN_AI_KEY" "SPEECH_KEY" "SPEECH_REGION")
+declare -a outputKeys=("OPEN_AI_ENDPOINT" "OPEN_AI_KEY" "SPEECH_KEY" "SPEECH_REGION" "SUFFEX")
 
 # Loop through each key and append to .env
 for keyName in "${outputKeys[@]}"; do
-    keyValue=$(terraform output -raw "$keyName")
-    echo "${keyName}=$keyValue" >> .env
+    keyValue=$(terraform output -raw $keyName)
+    if [[ $keyValue ]]; then
+        echo "${keyName^^}=\"$keyValue\"" >> $envFilePath
+    fi
 done
+
+python src/main.py
+
+# terraform destroy -auto-approve
